fixed progress bar and mkdir-ing

+++ run.py
@@ -1,14 +1,17 @@
 import argparse
-import tqdm
-from generate_composites import CompositeGenerator
+import json
 import os
 import os.path as osp
-import json
-import cv2
 import signal
+
+import cv2
+import tqdm
+from generate_composites import CompositeGenerator
 
 # Global Boolean variable that indicates that a signal has been received
 interrupted = False
+
+
 def signal_handler(signum, frame):
     global interrupted
     interrupted = True
@@ -21,19 +24,23 @@ def main():
     parser.add_argument("out_json_path")
     parser.add_argument("out_composites_dir")
     parser.add_argument("num_composites", type=int)
-    parser.add_argument("-p", "--prefix", default='')
+    parser.add_argument("-p", "--prefix")
     args = parser.parse_args()
 
     # Register the signal handler (for handling slurm job cancellation)
     signal.signal(signal.SIGTERM, signal_handler)
 
     # Prefix (if provided) to prepend to all generated image names and json name
-    prefix = args.prefix + "_" if args.prefix != "" else ""
+    if args.prefix is not None:
+        print("Prefix:", args.prefix)
+        prefix = args.prefix + "_"
+    else:
+        prefix = ""
     j_basename = osp.basename(args.out_json_path)
     json_path = args.out_json_path.replace(j_basename, prefix + j_basename)
 
     if not osp.isdir(args.out_composites_dir):
-        os.mkdir(args.out_composites_dir)
+        os.makedirs(args.out_composites_dir, exist_ok=True)
 
     # Resume from existing json if it exists
     if osp.isfile(json_path):
@@ -52,7 +59,9 @@ def main():
     )
 
     try:
-        for idx in tqdm.trange(num_existing, args.num_composites):
+        for idx in tqdm.trange(args.num_composites):
+            if idx < num_existing:
+                continue
             out_path = osp.join(args.out_composites_dir, prefix + f"{idx:06}.png")
             composite = cg.generate_composite(out_path)
             cv2.imwrite(out_path, composite)
